Fixed an issue where plot filter reduction was not applied on testnet10 if config was missing
